Revert "Lots of changes. Too many to describe well."
This reverts commit 7a0b8d2e7d697141cdbbeebc4d772437ec151e0a.
